resolve duo splitting edge case
